refactor: Merge types/docs from testing/auth/oauth_token_manager.pyi
- Merged type hints and docstrings from crudclient/testing/auth/oauth_token_manager.pyi into crudclient/testing/auth/oauth_token_manager.py.

- Deleted crudclient/testing/auth/oauth_token_manager.pyi.
