--- crudclient/testing/auth/oauth_token_manager.pyi
@@ -1,140 +1,0 @@
-"""
-OAuth token management utilities for testing.
-
-This module provides a class for managing OAuth tokens, including creation,
-validation, and refreshing of tokens.
-"""
-
-from datetime import datetime, timedelta
-from typing import Any, Dict, List, Mapping, Optional, Set, Union
-
-class OAuthTokenManager:
-    """
-    OAuth token manager for handling token lifecycle.
-
-    This class provides methods for creating, validating, and refreshing OAuth tokens,
-    as well as managing users for password grant type.
-    """
-
-    access_tokens: Dict[str, Dict]
-    refresh_tokens: Dict[str, str]
-    authorization_codes: Dict[str, Dict]
-    current_access_token: str
-    current_refresh_token: str
-    user_credentials: Dict[str, Dict]
-
-    def __init__(self) -> None:
-        """
-        Initialize an OAuth token manager.
-        """
-        ...
-
-    def initialize_default_token(self, client_id: str, scope: Optional[str]) -> None:
-        """
-        Initialize a default token for the OAuth server.
-
-        Args:
-            client_id: The client ID
-            scope: The scope for the token
-        """
-        ...
-
-    def create_token(
-        self,
-        client_id: str,
-        scope: Optional[str] = None,
-        expires_in: int = 3600,
-        token_type: str = "Bearer",
-        grant_type: str = "client_credentials",
-        user: Optional[str] = None,
-    ) -> Dict[str, Any]:
-        """
-        Create a new access token and refresh token pair.
-
-        Args:
-            client_id: The client ID
-            scope: The scope for the token
-            expires_in: The number of seconds until the token expires
-            token_type: The token type (e.g., "Bearer")
-            grant_type: The grant type used to obtain the token
-            user: The username for password grant type
-
-        Returns:
-            A dictionary with the token response
-        """
-        ...
-
-    def create_authorization_code(self, client_id: str, redirect_uri: str, scope: Optional[str] = None, state: Optional[str] = None) -> str:
-        """
-        Create a new authorization code.
-
-        Args:
-            client_id: The client ID
-            redirect_uri: The redirect URI
-            scope: The scope for the token
-            state: The state parameter
-
-        Returns:
-            The authorization code
-        """
-        ...
-
-    def validate_token(self, token: str) -> bool:
-        """
-        Check if a token is valid.
-
-        Args:
-            token: The token to validate
-
-        Returns:
-            True if the token is valid, False otherwise
-        """
-        ...
-
-    def refresh_token(self, refresh_token: str) -> Optional[Dict[str, Any]]:
-        """
-        Refresh an access token using a refresh token.
-
-        Args:
-            refresh_token: The refresh token
-
-        Returns:
-            A dictionary with the new token response, or None if the refresh token is invalid
-        """
-        ...
-
-    def revoke_token(self, token: str) -> bool:
-        """
-        Revoke an access token.
-
-        Args:
-            token: The token to revoke
-
-        Returns:
-            True if the token was revoked, False if it was not found
-        """
-        ...
-
-    def add_user(self, username: str, password: str, scopes: List[str]) -> None:
-        """
-        Add a user for password grant type.
-
-        Args:
-            username: The username
-            password: The password
-            scopes: The scopes for the user
-        """
-        ...
-
-    def validate_user(self, username: str, password: str) -> bool:
-        """
-        Validate user credentials.
-
-        Args:
-            username: The username
-            password: The password
-
-        Returns:
-            True if the credentials are valid, False otherwise
-        """
-        ...
